add test for dist with a note
Carmel's build artifacts do not have the dist/url info, so we might not know
if the artifact and dist/mirror/url combination actually match.

Also, Carmel will prefer newer versions of the artifacts if they satisfy the
requirements and onlhy use dist to resolve it for the first time.