disable static_install in Menlo. Fix #34
Dists with x_static_install: 1 opts in for the static_install, and ends up not
installing Module::Build::Tiny. This prevents the first run of carmel install to
not have the configure requirements specified in the Metadata.

Ideally, we can preserve this information in the snapshot so that Carmel
actually won't need MBT if all of the deps which would require MBT has
static_install. But until then, disable this feature.