auto upgrade core modules when it's pinned but is older than core
Major refactoring and change to the install command so it retries when new
dependencies are found that's pinned and also not found in the artifact. This
has been a major UX issue in the first run. install now retries while new
missing modules are found.

This also speeds up the install when all the satisfying artifacts are already
available, since previously it was resolving twice.